pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iab-toolkit"
version = "0.3.0"
description = "IAB Content Taxonomy v3.1 hybrid classifier using optimized embeddings and GPT"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "IAB Toolkit", email = "toolkit@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.3.7",
    "numpy",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
iab-hybrid = "iab_toolkit.hybrid.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["iab_toolkit*"]

[tool.setuptools.package-data]
iab_toolkit = ["hybrid/data/*"]
